moved mentions to a separate service
